Create new partition data type.
The original type was a list of integers indicating the
partition to which each ordinal of the points array belongs.
This is useless when the points array is resorted from y to x
order.
The new data structure is a vector of sets of pointers
to point elements.
All routines sort arrays of pointers, which are the used
for partitioning.
The number of induced points in a specific region defined
by an xy and y partition can be found efficiently O(n) as the
cardinality of the intersection of the two corresponding sets.